mascot-intercom-staging/test_reply.py
import asyncio
import os
from dotenv import load_dotenv
import httpx

load_dotenv()
TOKEN = os.getenv("INTERCOM_TOKEN")
ADMIN = os.getenv("INTERCOM_ADMIN_ID")

async def test():
    # Provide a hardcoded conversation_id from your logs, or I need to find one.
    # Wait, the logs show the error but not the conversation ID.
    pass
